refactor(sync): extract ManifestEntry to neutral internal/sync package
The sync HTTP handler (data-entry, TKT-PV0R3V) must reference the manifest DTO
without importing pgstore (which would pull pgx into the default build). Move
ManifestEntry to a backend-neutral internal/sync package; pgstore.ManifestSince
returns []sync.ManifestEntry. The consumer-side ManifestProvider interface will
live at the data-entry call site per CLAUDE.md. Archfile: register sync leaf,
pgstore mayDependOn sync.
