+++ .go-arch-lint.yml
@@ -65,6 +65,7 @@ components:
   migration:        { in: internal/migration }
   markdown:         { in: internal/markdown }
   canonical:        { in: internal/canonical }
+  sync:             { in: internal/sync }
   frontmatter:      { in: internal/frontmatter }
   filter:           { in: internal/filter }
   pattern:          { in: internal/pattern }
@@ -567,6 +568,7 @@ deps:
       - search
       - store
       - storeutil
+      - sync
     canUse:
       - pgx
   storeutil:
