+++ internal/store/pgstore/manifest.go
@@ -3,26 +3,9 @@ package pgstore
 import (
 	"context"
 	"fmt"
-)
 
-// ManifestEntry is one change in a sync manifest: a record that was created,
-// updated, or deleted since a cursor. The sync server (FEAT-NJ9FEN) turns a
-// slice of these into the wire manifest the client diffs against its index.
-//
-// Kind is "e" (entity) or "r" (relation). For an entity, IDA is the id and IDB/
-// IDC are empty; Typ is the entity type. For a relation, IDA/IDB/IDC are
-// from_id/rel_type/to_id and Typ is empty. Deleted is true when the entry comes
-// from a tombstone (the live row no longer exists). Seq is the change's position
-// in the global order — the highest Seq returned is the next cursor.
-type ManifestEntry struct {
-	Kind    string
-	IDA     string
-	IDB     string
-	IDC     string
-	Typ     string
-	Deleted bool
-	Seq     int64
-}
+	synctypes "github.com/Sourcehaven-BV/rela/internal/sync"
+)
 
 // ManifestSince returns every change with seq > cursor, in seq order: live
 // entity/relation rows (Deleted=false) UNION deletion tombstones (Deleted=true).
@@ -40,7 +23,7 @@ type ManifestEntry struct {
 // rather than rely on cursor 0 over a long-lived churny dataset. Tombstone
 // pruning (retention horizon) and manifest pagination (LIMIT + next-cursor) are
 // documented follow-ups (see TKT-GFJJ3S notes), not built here.
-func (s *Store) ManifestSince(ctx context.Context, cursor int64) ([]ManifestEntry, error) {
+func (s *Store) ManifestSince(ctx context.Context, cursor int64) ([]synctypes.ManifestEntry, error) {
 	const q = `
 		SELECT kind, a, b, c, typ, deleted, seq FROM (
 			SELECT 'e' AS kind, id AS a, '' AS b, '' AS c, type AS typ, false AS deleted, seq FROM entities
@@ -57,9 +40,9 @@ func (s *Store) ManifestSince(ctx context.Context, cursor int64) ([]ManifestEntr
 	}
 	defer rows.Close()
 
-	var out []ManifestEntry
+	var out []synctypes.ManifestEntry
 	for rows.Next() {
-		var e ManifestEntry
+		var e synctypes.ManifestEntry
 		if err := rows.Scan(&e.Kind, &e.IDA, &e.IDB, &e.IDC, &e.Typ, &e.Deleted, &e.Seq); err != nil {
 			return nil, fmt.Errorf("pgstore: manifest scan: %w", err)
 		}
